Add isLive property for live video
Removes scrubber (seek) bar and the file duration indicator when isLive
property is set to true. It’s default property set to false.
This solves the issue where when the player is playing a live stream,
if the user seeks, the app will crash. Plus a seek slider is not
desirable or functional with live video streams.